Create a new hoc and action file to store and handle functions related to allNetworks
