+++ src/actions/settings/networksActions.js
@@ -1,0 +1,31 @@
+import { createActions } from 'spunky';
+import { getStorage, setStorage } from '../../lib/storage';
+
+export const NETWORKS_ID = 'networks';
+
+// Setters
+export const setNetworks = createActions(NETWORKS_ID, (networks) => async () => {
+  await setStorage(NETWORKS_ID, networks);
+  return networks;
+});
+
+export const addNetwork = createActions(NETWORKS_ID, (network) => async () => {
+  const networks = await getStorage(NETWORKS_ID);
+  const newNetworks = [...networks, network];
+  await setStorage(NETWORKS_ID, newNetworks);
+  return newNetworks;
+});
+
+export const clearNetworks = createActions(NETWORKS_ID, () => async () => {
+  await setStorage(NETWORKS_ID, []);
+  return [];
+});
+
+// Getters
+export default createActions(NETWORKS_ID, () => async () => {
+  const allNetworks = await getStorage(NETWORKS_ID);
+  if (!allNetworks || !Array.isArray(allNetworks)) {
+    return [];
+  }
+  return allNetworks;
+});
